fix: extract ensureContrast to shared utility and apply to hero carousel
- Move ensureContrast() from game-timeline-card.tsx to lib/color-utils.ts
- Apply ensureContrast to hero-carousel.tsx badge text color
- Align badge border opacity to 40% in hero carousel (was 50%)
- Add guards for malformed hex and NaN values
